Initial commit of core `babel` package
The plan is to have packages `babel`, `babel-plugin`, and
`babel-runtime`.  The `babel` package wraps the `babel-core` npm
package, adding some tests of our own.  In the interest of less
verbose transpiled code, we will probably patch or hook into Babel
to make it generate code that runs against a small runtime support
library in `babel-runtime`.  `babel-plugin` will cause `.es` files to
be transpiled using an appropriate call to the `babel` package.

As we confirm which syntax features can be robustly transpiled and are
a positive contribution to Meteor style, we will enable them in the
plugin.